+++ Proyect/AutoLoads/Signals.gd
@@ -4,3 +4,8 @@ extends Node
 signal add_goal(team)
 signal on_goal()
 signal on_match_start()
+signal on_match_end()
+signal on_resume()
+signal on_pause()
+signal on_transition()
+signal start_versus()